+++ src/main/java/com/depromeet/sulsul/oauth2/filter/JwtAuthenticationFilter.java
@@ -4,10 +4,7 @@ import static org.springframework.http.HttpStatus.BAD_REQUEST;
 import static org.springframework.http.HttpStatus.OK;
 import static org.springframework.http.HttpStatus.UNAUTHORIZED;
 
-import com.depromeet.sulsul.common.response.dto.ResponseDto;
-import com.depromeet.sulsul.domain.member.repository.MemberRepository;
 import com.depromeet.sulsul.oauth2.provider.JwtTokenProvider;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.MalformedJwtException;
@@ -21,7 +18,6 @@ import javax.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpMethod;
-import org.springframework.http.HttpStatus;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
@@ -32,8 +28,8 @@ import org.springframework.web.filter.OncePerRequestFilter;
 @Component
 public class JwtAuthenticationFilter extends OncePerRequestFilter {
 
+  public static final String JWT_PARSE_RESULT = "jwt-parse-result";
   private final JwtTokenProvider jwtTokenProvider;
-  private final MemberRepository memberRepository;
 
   @Value("${client.url}")
   private String urlOfEnv;
@@ -46,9 +42,6 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
 
   @Value("${authentication.jwt.secretKey}")
   private String secretKey;
-
-  private ObjectMapper objectMapper = new ObjectMapper();
-
 
   @Override
   protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response,
@@ -71,12 +64,12 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
     try {
       Jwts.parserBuilder().setSigningKey(secretKey.getBytes()).build().parseClaimsJws(jwtToken);
     } catch (SignatureException | MalformedJwtException | UnsupportedJwtException | IllegalArgumentException e) {
-      setResponseJsonType(response);
-      processWithErrorResponseDto("[ERROR] jwt가 없거나 잘못 되었습니다.", BAD_REQUEST, response);
+      response.setIntHeader(JWT_PARSE_RESULT, BAD_REQUEST.value());
+      filterChain.doFilter(request, response);
       return;
     } catch (ExpiredJwtException e) {
-      setResponseJsonType(response);
-      processWithErrorResponseDto("[ERROR] jwt가 만료되었습니다.", UNAUTHORIZED, response);
+      response.setIntHeader(JWT_PARSE_RESULT, UNAUTHORIZED.value());
+      filterChain.doFilter(request, response);
       return;
     }
 
@@ -86,18 +79,5 @@ public class JwtAuthenticationFilter extends OncePerRequestFilter {
     SecurityContextHolder.getContext().setAuthentication(authentication);
 
     filterChain.doFilter(request, response);
-  }
-
-  private void processWithErrorResponseDto(String errorMessage, HttpStatus badRequest,
-      HttpServletResponse response) throws IOException {
-    String responseBodyWithJson = objectMapper.writeValueAsString(
-        ResponseDto.ERROR(errorMessage, badRequest));
-    response.getWriter().write(responseBodyWithJson);
-    response.setStatus(badRequest.value());
-  }
-
-  private void setResponseJsonType(HttpServletResponse response) {
-    response.setContentType("application/json");
-    response.setCharacterEncoding("utf-8");
   }
 }
